refactor(apps/bots): move hardcoded channel actions from bus listener to channel service
